+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- **Stoppage Time Display** - Goals in stoppage time now display properly (e.g., "45+2'")
 
 ### Changed